requirements.txt
absl-py==0.13.0
biopython==1.79
chex==0.0.7
dm-haiku==0.0.4
dm-tree==0.1.6
docker==5.0.0
immutabledict==2.0.0
jax==0.2.14
jmp==0.0.2
ml-collections==0.1.0
mpi4jax==0.3.2
mpi4py==3.1.1
numpy==1.19.5
scipy==1.7.0
tensorflow-cpu==2.5.0
